不同浏览器不能同时登录同一账号

+++ src/main/java/boot/spring/mapper/LoginMapper.java
@@ -4,5 +4,5 @@ import boot.spring.po.Staff;
 
 public interface LoginMapper {
 	Staff getpwdbyname(String name);
-	Staff getnamebyid(long id);
+	Staff getnamebyid(Integer id);
 }
