+++ src/main/java/boot/spring/po/User.java
@@ -1,7 +1,7 @@
 package boot.spring.po;
 
 public class User {
-	Long uid;
+    int uid;
 	
 	String name;
 	
@@ -9,26 +9,25 @@ public class User {
 		super();
 	}
 
-	public User(Long uid, String name) {
+	public User(int uid, String name) {
 		super();
 		this.uid = uid;
 		this.name = name;
 	}
 
-	public Long getUid() {
-		return uid;
-	}
+    public int getUid() {
+        return uid;
+    }
 
-	public void setUid(Long uid) {
-		this.uid = uid;
-	}
+    public void setUid(int uid) {
+        this.uid = uid;
+    }
 
-	public String getName() {
-		return name;
-	}
+    public String getName() {
+        return name;
+    }
 
-	public void setName(String name) {
-		this.name = name;
-	}
-	
+    public void setName(String name) {
+        this.name = name;
+    }
 }
